[TWINFACES-65] added general and options pages for datalists
