+++ src/app/datalists/[datalistId]/views/index.ts
@@ -1,0 +1,2 @@
+export * from "./datalist-general";
+export * from "./datalist-options";
